# docs/swagger.yaml
basePath: /
definitions:
...
  http.ErrorResponse:
    properties:
      message:
        description: Error message
        type: string
    type: object
  http.StatusResponse:
    properties:
      status:
        description: Status message
        type: string
    type: object
...
  table.CreateTableInput:
    properties:
      name:
        description: Table name
        type: string
      organization:
        description: Organization ID
        type: string
    required:
    - name
    - organization
    type: object
  table.Table:
    properties:
      id:
        description: Table ID
        type: string
      name:
        description: Table name
        type: string
      organization:
        description: Organization ID
        type: string
    required:
    - name
    - organization
    type: object
  table.UpdateTableInput:
    properties:
      id:
        description: Table ID
        type: string
      name:
        description: Table name
        type: string
      organisation:
        description: Organization ID
        type: string
    type: object
...
info:
  contact:
    email: evgeniydammer@gmail.com
    name: Evgeniy Dammer
  description: marketplace API service
  title: marketplace API
  version: "1.0"
paths:
...
  /api/v1/tables/:
    patch:
      consumes:
      - application/json
      description: Update table method.
      parameters:
      - description: Table data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/table.UpdateTableInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Update table method.
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: Create table method.
      parameters:
      - description: Table data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/table.CreateTableInput'
      produces:
      - application/json
      responses:
        "200":
          description: Table ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Create table method.
      tags:
      - tables
  /api/v1/tables/{org_id}:
    get:
      consumes:
      - application/json
      description: Get all tables method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Table List
          schema:
            items:
              $ref: '#/definitions/table.Table'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all tables method.
      tags:
      - tables
  /api/v1/tables/{org_id}/{id}:
    delete:
      consumes:
      - application/json
      description: Delete table method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete table method.
      tags:
      - tables
    get:
      consumes:
      - application/json
      description: Get table by id method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Table data
          schema:
            $ref: '#/definitions/table.Table'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get table by id method.
      tags:
      - tables
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"